feat: cập nhật lại quy tắc đặt tên bảng.
- Danh sách api:
    - Tạo, thay đổi, nhận thông tin sản phẩm.
    - Tạo, thay đổi, xóa, nhận giá sản phẩm.
    - Tạo, thay đổi, nhận kích thước cho sản phẩm.
    - Tạo, thay đổi, nhận danh mục sản phẩm.
    - Tạo, thay đổi, nhận đơn vị tính.
    - Tạo, thay đổi, nhận kích thước sản phẩm.
    - Tạo, thay đổi, nhận thông tin tài khoản.
    - Tạo, thay đổi, nhận thông tin vai trò.
    - Tạo, thay đổi, nhận thông tin người quản lý.
    - Tạo, thay đổi, nhận thông tin nhân viên.
    - Tạo, thay đổi, nhận thông tin khách hàng.
    - Tạo, thay đổi, nhận thông tin khách hàng.
    - Tạo, thay đổi, nhận thông tin chương trình thành viên.
    - Tạo, thay đổi, nhận thông tin khu vực.
    - Tạo, thay đổi, nhận thông tin bàn.
    - Tạo, thay đổi, nhận thông tin cửa hàng.
    - Tạo, thay đổi, nhận thông tin mã khuyến mãi.
    - Tạo, thay đổi, nhận thông tin chương trình khuyến mãi.
    - Tạo đơn hàng, hủy đơn hàng.
    - Tiến hành thanh toán:
        - Thanh toán bằng tiền mặt.

Author: Nguyen Ngoc Phu

+++ backend/application/src/main/java/com/mts/backend/application/payment/response/PaymentReportResponse.java
@@ -17,6 +17,6 @@ import org.springframework.data.domain.Page;
 public class PaymentReportResponse {
     private BigDecimal totalAmount;
     private BigDecimal averageAmount;
-    private Integer totalPayment;
+    private Long totalPayment;
     private List<PaymentDetailResponse> paymentDetailResponses;
 }
